Removiendo la logica de room para implementarla despues
